# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JakeAI Platform",
    "description": "Enterprise-grade embedded AI platform integrating with FinnApiGo as upstream Identity, Authorization, and Business Core.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/gateway/chat/completions": {
      "post": {
        "tags": [
          "AI Gateway"
        ],
        "summary": "Proxy Chat Completions",
        "description": "OpenAI-compatible inference proxy with Tier 1 Redis exact caching and quota deduction.",
        "operationId": "proxy_chat_completions_api_v1_gateway_chat_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewayChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "FinnApiGoAuth": []
          }
        ]
      }
    },
    "/api/v1/gateway/models": {
      "get": {
        "tags": [
          "AI Gateway"
        ],
        "summary": "List Available Models",
        "description": "Return catalog of available models supported by the AI Gateway.",
        "operationId": "list_available_models_api_v1_gateway_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "FinnApiGoAuth": []
          }
        ]
      }
    },
    "/api/v1/gateway/quotas": {
      "get": {
        "tags": [
          "AI Gateway"
        ],
        "summary": "Get Tenant Quota",
        "description": "Retrieve active period token usage, limits, and budget health.",
        "operationId": "get_tenant_quota_api_v1_gateway_quotas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaStatus"
                }
              }
            }
          }
        },
        "security": [
          {
            "FinnApiGoAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "AI Gateway"
        ],
        "summary": "Update Tenant Quota",
        "description": "Configure monthly token quota limit for a tenant.",
        "operationId": "update_tenant_quota_api_v1_gateway_quotas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuotaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "FinnApiGoAuth": []
          }
        ]
      }
    },
...
    "/api/v1/coding/tool-result": {
      "post": {
        "tags": [
          "Coding Agent & Tool Bridge"
        ],
        "summary": "Submit Tool Result",
        "description": "Submit local developer workstation tool execution outcome and resume agent workflow.",
        "operationId": "submit_tool_result_api_v1_coding_tool_result_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolResultSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumedExecutionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "FinnApiGoAuth": []
          }
        ]
      }
    },
    "/api/v1/coding/resume": {
      "post": {
        "tags": [
          "Coding Agent & Tool Bridge",
          "Internal Coding Bridge"
        ],
        "summary": "Internal Resume Endpoint",
        "description": "Internal perimeter resume bridge invoked by FinnApiGo with Invariant 4 authentication.",
        "operationId": "internal_resume_endpoint_api_v1_coding_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InternalResumeSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumedExecutionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "tags": [
          "OpenAI Compatible AI Gateway"
        ],
        "summary": "Proxy Chat Completions",
        "description": "OpenAI-compatible inference proxy with Tier 1 Redis exact caching and quota deduction.",
        "operationId": "proxy_chat_completions_v1_chat_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewayChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "FinnApiGoAuth": []
          }
        ]
      }
    },
    "/v1/models": {
      "get": {
        "tags": [
          "OpenAI Compatible AI Gateway"
        ],
        "summary": "List Available Models",
        "description": "Return catalog of available models supported by the AI Gateway.",
        "operationId": "list_available_models_v1_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "FinnApiGoAuth": []
          }
        ]
      }
    },
    "/v1/quotas": {
      "get": {
        "tags": [
          "OpenAI Compatible AI Gateway"
        ],
        "summary": "Get Tenant Quota",
        "description": "Retrieve active period token usage, limits, and budget health.",
        "operationId": "get_tenant_quota_v1_quotas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaStatus"
                }
              }
            }
          }
        },
        "security": [
          {
            "FinnApiGoAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "OpenAI Compatible AI Gateway"
        ],
        "summary": "Update Tenant Quota",
        "description": "Configure monthly token quota limit for a tenant.",
        "operationId": "update_tenant_quota_v1_quotas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuotaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "FinnApiGoAuth": []
          }
        ]
      }
    },
    "/internal/v1/coding/tool-result": {
      "post": {
        "tags": [
          "Internal Coding Bridge"
        ],
        "summary": "Submit Tool Result",
        "description": "Submit local developer workstation tool execution outcome and resume agent workflow.",
        "operationId": "submit_tool_result_internal_v1_coding_tool_result_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolResultSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumedExecutionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "FinnApiGoAuth": []
          }
        ]
      }
    },
    "/internal/v1/coding/resume": {
      "post": {
        "tags": [
          "Internal Coding Bridge",
          "Internal Coding Bridge"
        ],
        "summary": "Internal Resume Endpoint",
        "description": "Internal perimeter resume bridge invoked by FinnApiGo with Invariant 4 authentication.",
        "operationId": "internal_resume_endpoint_internal_v1_coding_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InternalResumeSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumedExecutionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChatMessage": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role",
            "description": "role: system, user, assistant"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "message content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage",
        "description": "OpenAI-compatible message format."
      },
...
      "GatewayChatRequest": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "description": "Target model name",
            "default": "gemini-1.5-flash"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "minItems": 1,
            "title": "Messages",
            "description": "List of messages"
          },
          "temperature": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "default": 0.7
          },
          "max_tokens": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Tokens",
            "default": 1024
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "GatewayChatRequest",
        "description": "OpenAI-compatible chat completions proxy request."
      },
      "GatewayChatResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "chat.completion"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "choices": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Choices"
          },
          "usage": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Usage"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached"
          },
          "tokens_saved": {
            "type": "integer",
            "title": "Tokens Saved"
          },
          "reduction_percentage": {
            "type": "number",
            "title": "Reduction Percentage",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "id",
          "created",
          "model",
          "choices",
          "usage",
          "cached",
          "tokens_saved"
        ],
        "title": "GatewayChatResponse",
        "description": "OpenAI-compatible chat completion response with JakeAI caching metadata."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InternalResumeSubmission": {
        "properties": {
          "call_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Call Id"
          },
          "tenant_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Tenant Id"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          }
        },
        "type": "object",
        "required": [
          "call_id",
          "tenant_id",
          "result"
        ],
        "title": "InternalResumeSubmission",
        "description": "Payload routed internally by FinnApiGo edge gateway."
      },
      "ModelItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "model"
          },
          "created": {
            "type": "integer",
            "title": "Created",
            "default": 1700000000
          },
          "owned_by": {
            "type": "string",
            "title": "Owned By",
            "default": "jakeai"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "ModelItem",
        "description": "OpenAI model descriptor."
      },
      "ModelListResponse": {
        "properties": {
          "object": {
            "type": "string",
            "title": "Object",
            "default": "list"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ModelItem"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ModelListResponse",
        "description": "OpenAI-compatible model list response."
      },
...
      "QuotaStatus": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "quota_limit": {
            "type": "integer",
            "title": "Quota Limit"
          },
          "tokens_used": {
            "type": "integer",
            "title": "Tokens Used"
          },
          "tokens_remaining": {
            "type": "integer",
            "title": "Tokens Remaining"
          },
          "percentage_used": {
            "type": "number",
            "title": "Percentage Used"
          },
          "is_suspended": {
            "type": "boolean",
            "title": "Is Suspended"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "quota_limit",
          "tokens_used",
          "tokens_remaining",
          "percentage_used",
          "is_suspended",
          "warning"
        ],
        "title": "QuotaStatus",
        "description": "Current tenant quota consumption and status."
      },
...
      "ResumedExecutionResult": {
        "properties": {
          "call_id": {
            "type": "string",
            "title": "Call Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "resumed_at": {
            "type": "number",
            "title": "Resumed At"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "tool_acknowledged": {
            "type": "boolean",
            "title": "Tool Acknowledged"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "call_id",
          "status",
          "resumed_at",
          "tenant_id",
          "message",
          "tool_acknowledged"
        ],
        "title": "ResumedExecutionResult",
        "description": "Outcome of rehydrated execution after tool result processing."
      },
...
      "ToolResultSubmission": {
        "properties": {
          "call_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Call Id",
            "description": "Unique tool call execution identifier generated during interrupt()"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result",
            "description": "Structured tool output (e.g. file contents, AST matches, test outcome)"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id",
            "description": "Optional LangGraph thread identifier"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id",
            "description": "Optional conversation identifier"
          }
        },
        "type": "object",
        "required": [
          "call_id",
          "result"
        ],
        "title": "ToolResultSubmission",
        "description": "Payload submitted by client application upon completing a local tool execution."
      },
      "UpdateQuotaRequest": {
        "properties": {
          "new_limit": {
            "type": "integer",
            "minimum": 10000.0,
            "title": "New Limit",
            "description": "New token limit (e.g. 5000000)"
          }
        },
        "type": "object",
        "required": [
          "new_limit"
        ],
        "title": "UpdateQuotaRequest",
        "description": "Payload to update tenant monthly token quota."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "FinnApiGoAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "FinnApiGo asymmetric RS256/HS256 JWT access token. Must contain 'sub', 'tenant_id', and authorized roles/scopes."
      }
    }
  },
...
  "security": [
    {
      "FinnApiGoAuth": []
    }
  ]
}